Adapted cairo addons to latest bill 
source refactoring and documenting. (great work bill!)

Cairo intergration not finished 
but already working better than it was
plus now working on win32 and mac os x platforms.

This version was tested under mac osx
will check other platforms in few minutes...



git-svn-id: file:///fltk/svn/trunk@5237 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fltk/draw.h
@@ -103,7 +103,7 @@ inline Flags drawflags() {return drawflags_;}
 inline Flags drawflags(Flags f) {return drawflags_ & f;}
 
 // line type:
-FL_API void line_style(int, int width=0, char* dashes=0);
+FL_API void line_style(int, double width=0.0, char* dashes=0);
 enum {
   SOLID	= 0,
   DASH	= 1,
